requirements.txt
streamlit
langchain
langgraph
langchain-groq
langchain-chroma
duckduckgo-search
reportlab
python-dotenv
langchain-huggingface